+++ client/src/assets/locales/nb/translation.json
@@ -13,7 +13,8 @@
   "bestillinger.loggInnForÅSeBestillinger": "Logg inn for å se dine innsendte bestillinger.",
   "bestillinger.bestillingTil": "Bestilling til {{navn}}",
   "bestillinger.bestill": "Bestill",
-   "bestillinger.bestilldel": "Bestill del",
+  "bestillinger.bestilldel": "Bestill del",
+  "bestillinger.henterDeler": "Henter deler...",
   "bestillinger.del.ikkePåLager": "Delen var ikke på lager og kan ha lenger leveringstid.",
   "bestillinger.tilbakeTilBestillingen": "Tilbake til bestillingen",
   "bestillinger.bestillDelerTil": "Bestill deler til {{navn}}",
